Refactor log messages.
Currently, there are four types of log messages generated by the `refresh.py`
script:

    - error messages (in red)
    - warning messages (in yellow)
    - info messages (in blue)
    - success messages (in green)

Previously, logging calls were scattered throughout the script.

Now, logging calls are refactored into `log_*` functions, and (a few) SGR
control sequences [^1] live in an `enum` so the control sequences can have
symbolic names (which is incredibly helpful, unless you enjoy memorizing whether
`\033[0;34m` is blue or not).

The `log_*` functions have the same interface as `print`, for easy parity and
backwards compatibility, but given that no invocations currently use `sep`,
`end`, or `flush`, it would be possible to adapt an interface closer to that of
the `logging` module.

This commit does _not_ introduce alternative colors (e.g. bold red instead of
normal red), nor does it check whether the file to which log messages are
written is even a TTY.

[^1]: https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_(Select_Graphic_Rendition)_parameters
[^2]: https://docs.python.org/3/library/logging.html