Use `sys.exit` instead of `site`'s exit.
The familiar `exit` and `quit` we sometimes use at the interpreter are not, in
fact, Python built-in functions [^1], and are instead part of the `site` [^2]
module's side effects, which are typically (but not necessarily!) loaded as part
of Python's initialization.

The `site` module is bananas, and is responsible for a lot of Python's "magic"
startup behavior, but `site.setquit` in the CPython source [^3] shows that
`exit` and `quit` need to be explicitly added before they can be used.

    $ python -c "print('exit' in __builtins__.__dict__)"
    True
    $ python -S -c "print('exit' in __builtins__.__dict__)"
    False
    $ python -S -c "import site; site.main(); print('exit' in __builtins__.__dict__)"
    True

We already import `sys`, so it's easier to just call `sys.exit`. After all,
`_sitebuiltins.Quitter` [^4] just raises `SystemExit` after some indirection.

Adam Johnson has a friendly writeup [^5] on the many ways to exit in Python.

[^1]: https://docs.python.org/3/library/functions.html
[^2]: https://docs.python.org/3/library/site.html
[^3]: https://github.com/python/cpython/blob/main/Lib/site.py
[^4]: https://github.com/python/cpython/blob/main/Lib/_sitebuiltins.py
[^5]: https://adamj.eu/tech/2021/10/10/the-many-ways-to-exit-in-python/