Add reading times to READMEs.
Estimated reading times are valuable because prospective readers can preallocate
contiguous mental capacity to understand a document before the begin reading.

Estimated reading times can be problematic when (1) they easily get out of sync
with highly volatile content, and (2) they dissuade readers from even starting
to read, neither of which are true here because (1) this tool is stable enough
that the READMEs change only infrequently, and are unlikely to change in a way
that affects the overall read time, and (2) 15m is not an overwhelming amount of
time for a document.