Fix: Use os.readlink instead of Path.readlink before Python 3.9.
`Path.readlink` is not available before Python 3.9, which is causing the tool to complain in a project I'm working on. This code was implemented as `os.readlink()` before rules_python was introduced (per https://github.com/hedronvision/bazel-compile-commands-extractor/commit/0e5b1aa26d87a431d2a52676d0b9ce469448ee54), but didn't seem to be reverted when rules_python was reverted.

Verified locally that this fixes the issue for me.